+++ dlt_openapi/detector/default/__init__.py
@@ -2,10 +2,11 @@
 Default open source detector
 """
 
+import json
 from typing import Dict, List, Optional, Tuple, Union, cast
 
 from dlt_openapi.config import Config
-from dlt_openapi.detector.base_detector import BaseDetector
+from dlt_openapi.detector.base_detector import GLOBAL_WARNING_KEY, BaseDetector
 from dlt_openapi.detector.default import utils
 from dlt_openapi.detector.default.primary_key import detect_primary_key_by_name
 from dlt_openapi.parser.endpoints import Endpoint, EndpointCollection, Response, TransformerSetting
@@ -36,19 +37,36 @@ from .const import (
     RE_UNIQUE_KEY,
 )
 from .utils import to_int
+from .warnings import (
+    BaseDetectionWarning,
+    DataResponseNoBodyWarning,
+    DataResponseUndetectedWarning,
+    PossiblePaginatorWarning,
+    PrimaryKeyNotFoundWarning,
+    UnresolvedPathParametersWarning,
+    UnsupportedSecuritySchemeWarning,
+)
 
 Tree = Dict[str, Union["str", "Tree"]]
 
 
 class DefaultDetector(BaseDetector):
+
+    warnings: Dict[str, List[BaseDetectionWarning]] = {}
+
     def __init__(self, config: Config) -> None:
         self.config = config
 
     def run(self, open_api: OpenapiParser) -> None:
         """Run the detector"""
+        self.warnings = {}
+
+        # detect security stuff
+        self.detect_security_schemes(open_api)
 
         # discover stuff from responses
         self.detect_paginators_and_responses(open_api.endpoints)
+        self.detect_global_pagination(open_api)
 
         # discover parent child relationship
         self.detect_parent_child_relationships(open_api.endpoints)
@@ -61,6 +79,46 @@ class DefaultDetector(BaseDetector):
 
         # and sort resources by table name
         open_api.endpoints.endpoints.sort(key=lambda e: e.detected_table_name)
+
+        # add some warnings
+        for e in open_api.endpoints.endpoints:
+            if params := e.unresolvable_path_param_names:
+                self._add_warning(UnresolvedPathParametersWarning(params), e)
+
+    def detect_security_schemes(self, open_api: OpenapiParser) -> None:
+        schemes = list(open_api.security_schemes.values())
+
+        # detect scheme settings
+        # TODO: make this a bit nicer
+        for scheme in schemes:
+
+            if scheme.type == "apiKey":
+                scheme.detected_secret_name = "api_key"
+                scheme.detected_auth_vars = f"""
+            "type": "api_key",
+            "api_key": api_key,
+            "name": "{scheme.name}",
+            "location": "{scheme.location}"
+"""
+            elif scheme.type == "http" and scheme.scheme == "basic":
+                scheme.detected_secret_name = "password"
+                scheme.detected_auth_vars = """
+            "type": "http_basic",
+            "username": "username",
+            "password": password,
+"""
+            elif scheme.type == "http" and scheme.scheme == "bearer":
+                scheme.detected_secret_name = "token"
+                scheme.detected_auth_vars = """
+            "type": "bearer",
+            "token": token,
+"""
+
+        # find default scheme
+        if len(schemes) and schemes[0].supported:
+            open_api.detected_global_security_scheme = schemes[0]
+        elif len(schemes) and not schemes[0].supported:
+            self._add_warning(UnsupportedSecuritySchemeWarning(schemes[0].name))
 
     def detect_resource_names(self, endpoints: EndpointCollection) -> None:
         """iterate all endpoints and find a strategy to select the right resource name"""
@@ -144,9 +202,43 @@ class DefaultDetector(BaseDetector):
                 endpoint.detected_data_response.detected_payload = self.detect_response_payload(
                     endpoint.detected_data_response, expect_list=expect_list
                 )
-                self.detect_primary_key(endpoint.detected_data_response, endpoint.path)
+                self.detect_primary_key(endpoint, endpoint.detected_data_response, endpoint.path)
 
-    def detect_primary_key(self, response: Response, path: str) -> None:
+    def detect_global_pagination(self, open_api: OpenapiParser) -> None:
+        """go through all detected paginators and see which one we can set as global"""
+        paginator_by_key: Dict[str, Pagination] = {}
+        paginator_count: Dict[str, int] = {}
+
+        # count how many every paginator appears
+        for endpoint in open_api.endpoints.endpoints:
+            if not endpoint.detected_pagination:
+                continue
+            params = endpoint.detected_pagination.paginator_config
+            key = json.dumps(params, sort_keys=True)
+            paginator_by_key[key] = endpoint.detected_pagination
+            paginator_count.setdefault(key, 0)
+            paginator_count[key] += 1
+
+        # no paginators found
+        if len(paginator_by_key) == 0:
+            return
+
+        # sort dict by value descending, so most used paginator is at the top
+        sorted_paginator_count = sorted(paginator_count.items(), key=lambda item: item[1] * -1)
+
+        # we only set a global paginator, if we found one paginator, or if the top paginator has
+        # a higher count than the second most used one
+        if not (len(paginator_by_key) == 1 or sorted_paginator_count[0][1] > sorted_paginator_count[1][1]):
+            return
+
+        global_paginator = paginator_by_key[sorted_paginator_count[0][0]]
+
+        # set global paginator on base object but also set on all endpoints
+        open_api.detected_global_pagination = global_paginator
+        for e in open_api.endpoints.endpoints:
+            e.detected_global_pagination = global_paginator
+
+    def detect_primary_key(self, e: Endpoint, response: Response, path: str) -> None:
         """detect the primary key from the payload"""
         if not response.detected_payload:
             return
@@ -179,6 +271,9 @@ class DefaultDetector(BaseDetector):
         elif uuid_paths:
             response.detected_primary_key = uuid_paths[0]
 
+        if not response.detected_primary_key:
+            self._add_warning(PrimaryKeyNotFoundWarning(), e)
+
     def detect_main_response(self, endpoint: Endpoint) -> Optional[Response]:
         """Get main response and pagination for endpoint"""
 
@@ -190,6 +285,12 @@ class DefaultDetector(BaseDetector):
                 break  # this will always be the right one
             if response.status_code.startswith("2") and not main_response:
                 main_response = response
+
+        if not main_response:
+            self._add_warning(DataResponseUndetectedWarning(), endpoint)
+
+        if main_response and not main_response.schema:
+            self._add_warning(DataResponseNoBodyWarning(), endpoint)
 
         return main_response
 
@@ -354,6 +455,10 @@ class DefaultDetector(BaseDetector):
         #
         # Nothing found :(
         #
+        pagination_params = [*cursor_params, *offset_params, *limit_params, *page_params]
+        if pagination_params:
+            self._add_warning(PossiblePaginatorWarning([p.name for p in pagination_params]), endpoint)
+
         return None
 
     def detect_parent_child_relationships(self, endpoints: EndpointCollection) -> None:
@@ -397,3 +502,11 @@ class DefaultDetector(BaseDetector):
             endpoint.detected_parent = find_nearest_list_parent(endpoint)
             if endpoint.detected_parent:
                 endpoint.detected_parent.detected_children.append(endpoint)
+
+    def get_warnings(self) -> Dict[str, List[BaseDetectionWarning]]:
+        return self.warnings
+
+    def _add_warning(self, warning: BaseDetectionWarning, e: Optional[Endpoint] = None) -> None:
+        key = e.id if e else GLOBAL_WARNING_KEY
+        warning_list = self.warnings.setdefault(key, [])
+        warning_list.append(warning)