+++ tests/integration/basics/test_data_selector.py
@@ -2,12 +2,15 @@ from typing import Any, Dict
 
 import pytest
 
+from dlt_openapi.config import Config
 from tests.integration.utils import get_indexed_resources
 
 
 @pytest.fixture(scope="module")
 def data_selectors() -> Dict[str, str]:
-    resources = get_indexed_resources("artificial", "data_selector.yml", name_resources_by_operation=True)
+    resources = get_indexed_resources(
+        "artificial", "data_selector.yml", config=Config(name_resources_by_operation=True)
+    )
     return {name: resource.get("endpoint").get("data_selector") for name, resource in resources.items()}  # type: ignore