+++ tests/integration/utils.py
@@ -6,7 +6,7 @@ from typing import Dict, Iterable, List, Literal, Union, cast
 from dlt.common.validation import validate_dict
 from dlt.extract.source import DltSource
 
-from dlt_openapi import _get_project_for_url_or_path
+from dlt_openapi import Project, _get_project_for_url_or_path
 from dlt_openapi.config import Config
 from dlt_openapi.renderer.default import REST_API_SOURCE_LOCATION
 from sources.sources.rest_api.typing import EndpointResource, RESTAPIConfig
@@ -17,11 +17,27 @@ LOCAL_DIR = "tests/_local/"
 TType = Literal["artificial", "original", "extracted"]
 
 
+def get_detected_project_from_open_api(case: str, config: Config) -> Project:
+
+    config = config or Config()
+
+    config.project_name = "test"
+    config.package_name = "test"
+    config.prepare()
+
+    # get project and render source into string
+    project = _get_project_for_url_or_path(url=None, path=case, config=config)  # type: ignore
+    project.parse()
+    project.detect()
+
+    return project
+
+
 def get_source_or_dict_from_open_api(
     case: str,
     rt: Literal["source", "dict"] = "source",
     base_url: str = "base_url",
-    name_resources_by_operation: bool = True,
+    config: Config = None,
 ) -> Union[DltSource, RESTAPIConfig]:
     """
     This function renders the source into a string and returns the extracted
@@ -36,12 +52,7 @@ from typing import Any
 Oauth20Credentials = Any
 """
 
-    config = Config(name_resources_by_operation=name_resources_by_operation, project_name="test", package_name="test")
-
-    # get project and render source into string
-    project = _get_project_for_url_or_path(url=None, path=case, config=config)  # type: ignore
-    project.parse()
-    project.detect()
+    project = get_detected_project_from_open_api(case, config)
     project.render(dry=True)
     source = project.renderer._render_source()  # type: ignore
 
@@ -67,27 +78,26 @@ Oauth20Credentials = Any
     os.environ["API_KEY"] = "some api key"
 
     module = importlib.import_module(local.replace("/", "."))
+    importlib.reload(module)
 
     remove_tree(LOCAL_DIR + "rest_api")
 
     return cast(DltSource, module.test_source())
 
 
-def get_source(case: str, base_url: str = "base_url", name_resources_by_operation: bool = False) -> DltSource:
+def get_source(case: str, base_url: str = "base_url", config: Config = None) -> DltSource:
     """Creates source for case path and returns it"""
     return cast(
         DltSource,
-        get_source_or_dict_from_open_api(
-            case, "source", base_url, name_resources_by_operation=name_resources_by_operation
-        ),
+        get_source_or_dict_from_open_api(case, "source", base_url, config),
     )
 
 
-def get_dict_by_path(case: str, validate: bool = True, name_resources_by_operation: bool = False) -> RESTAPIConfig:
+def get_dict_by_path(case: str, validate: bool = True, config: Config = None) -> RESTAPIConfig:
     """Renders dict for case path and returns it"""
     api_dict = cast(
         RESTAPIConfig,
-        get_source_or_dict_from_open_api(case, "dict", name_resources_by_operation=name_resources_by_operation),
+        get_source_or_dict_from_open_api(case, "dict", config=config),
     )
     api_dict["client"]["base_url"] = "something"
     if validate:
@@ -95,19 +105,20 @@ def get_dict_by_path(case: str, validate: bool = True, name_resources_by_operati
     return api_dict
 
 
-def get_dict_by_case(
-    type: TType, case: str, validate: bool = True, name_resources_by_operation: bool = False
-) -> RESTAPIConfig:
+def get_dict_by_case(type: TType, case: str, validate: bool = True, config: Config = None) -> RESTAPIConfig:
     """Renders dict for case path and returns it"""
     path = case_path(type, case)
-    return get_dict_by_path(path, validate=validate, name_resources_by_operation=name_resources_by_operation)
+    return get_dict_by_path(path, validate=validate, config=config)
 
 
-def get_indexed_resources(
-    type: TType, case: str, name_resources_by_operation: bool = False
-) -> Dict[str, EndpointResource]:
+def get_project_by_case(type: TType, case: str, config: Config = None) -> Project:
+    path = case_path(type, case)
+    return get_detected_project_from_open_api(path, config)
+
+
+def get_indexed_resources(type: TType, case: str, config: Config = None) -> Dict[str, EndpointResource]:
     """get all found resources indexed by name"""
-    rendered_dict = get_dict_by_case(type, case, name_resources_by_operation=name_resources_by_operation)
+    rendered_dict = get_dict_by_case(type, case, config=config)
     return {entry["name"]: entry for entry in rendered_dict["resources"]}  # type: ignore